Merged PR 3328: Merge fix/PRI-6647 to develop
Insert post file details using a bulk insert. Use modified CoreRepositoryBase.BulkInsert method that will persist the assigned IDs in the database because the IDs are needed later in the request to calculate impressions. The method in CoreRepositoryBase does use the assigned IDs in the database because the SqlBulkCopy doesn't use the SqlBulkCopyOptions.KeepIdentity option.