Merged PR 3553: Merge fix/PRI-8493 to develop
When editing post file settings, change the route to not return all the file details and impressions. Instead it will only return the settings data that are needed.

Also changed the repo method to delete impressions to use a stored procedure because the previous ef query was running out of memory for larger files.