Merged PR 5280: PRI-14085: Updated the LogHelper to log Message object as Json instead of
PRI-14085: Updated the LogHelper to log Message object as Json instead of string, so that it's more readable and indexable.

Previous to this fix, the log was as below,

{"date":"2019-09-06T14:44:06.3553940-04:00","level":"INFO","thread":"1","logger":"Tam.Maestro.Common.Utilities.Logging.LogHelper",**"message":"{\"environment\":\"DEV\",\"serviceName\":\"BroadcastController\",\"message\":\"Initializing Broadcast Web Application.\",\"userName\":\"kmoorthy\",\"action\":\"Service is started\"}**"}

With the below fix, the log will be as below. (Message object is no longer escaped string and it's a Json object now and will be indexable in Chaos Search

{"date":"2019-09-10T15:19:31.5650523-04:00","level":"INFO","thread":"1","logger":"Tam.Maestro.Common.Utilities.Logging.LogHelper",**"message":{"environment":"DEV","serviceName":"BroadcastController","message":"Initializing Broadcast Web Application.","userName":"kmoorthy","action":"Service is started"}**}